requirements.txt
async-timeout==4.0.2
black==22.3.0
boto3==1.24.10
botocore==1.27.10
click==8.1.3
Deprecated==1.2.13
greenlet==1.1.2
jmespath==1.0.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
redis==4.3.3
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==1.4.37
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
wrapt==1.14.1
